Use fixed-format for plot axis labels; add new model and file picker delegate for met data; update Scintilla from 4.2.0 to 4.3.0 to improve folding performance.

+++ src/BufferZoneModel.cpp
@@ -109,7 +109,7 @@ bool BufferZoneModel::setData(const QModelIndex &index, const QVariant &value, i
                 it->distance = value.toDouble();
                 break;
             case 3:
-                it->duration = value.toDouble();
+                it->duration = value.toInt();
                 break;
             default:
                 return false;
